# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "QueueEntryStatus": {
                "enum": ["Downloading", "Queued"],
                "type": "string"
            },
            "DownloadDetails": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "progress": {
                        "type": "number",
                        "format": "double"
                    },
                    "size": {
                        "type": "number",
                        "format": "double"
                    },
                    "sizeLeft": {
                        "type": "number",
                        "format": "double"
                    },
                    "speed": {
                        "type": "number",
                        "format": "double"
                    },
                    "eta": {
                        "type": "string"
                    },
                    "start": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "additionalProperties": true
            },
            "VideoType": {
                "enum": ["TV", "MOVIE"],
                "type": "string"
            },
            "QueueEntryDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "pid": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/QueueEntryStatus"
                    },
                    "details": {
                        "$ref": "#/components/schemas/DownloadDetails"
                    },
                    "nzbName": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/VideoType"
                    },
                    "appId": {
                        "type": "string"
                    }
                },
                "required": ["id", "pid", "status", "nzbName", "type"],
                "type": "object",
                "additionalProperties": true
            },
            "EpisodeCacheDefinition": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cacheRefreshed": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": ["id", "url", "name"],
                "type": "object",
                "additionalProperties": true
            },
            "Record_string.string_": {
                "properties": {},
                "additionalProperties": {
                    "type": "string"
                },
                "type": "object",
                "description": "Construct a type with a set of properties K of type T"
            },
...
            "AppType": {
                "enum": ["SONARR", "RADARR", "PROWLARR", "SABNZBD", "NZBGET"],
                "type": "string"
            },
            "App": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AppType"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "api_key": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "number",
                        "format": "double"
                    },
                    "iplayarr": {
                        "properties": {
                            "useSSL": {
                                "type": "boolean"
                            },
                            "port": {
                                "type": "number",
                                "format": "double"
                            },
                            "host": {
                                "type": "string"
                            }
                        },
                        "required": ["useSSL", "port", "host"],
                        "type": "object"
                    },
                    "download_client": {
                        "properties": {
                            "port": {
                                "type": "number",
                                "format": "double"
                            },
                            "api_key": {
                                "type": "string"
                            },
                            "host": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "id": {
                                "type": "number",
                                "format": "double"
                            }
                        },
                        "required": ["id"],
                        "type": "object"
                    },
                    "indexer": {
                        "properties": {
                            "priority": {
                                "type": "number",
                                "format": "double"
                            },
                            "api_key": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "id": {
                                "type": "number",
                                "format": "double"
                            }
                        },
                        "required": ["priority", "id"],
                        "type": "object"
                    }
                },
                "required": ["id", "type", "name", "url", "iplayarr"],
                "type": "object",
                "additionalProperties": true
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
    "info": {
        "title": "iplayarr",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
...
        "/api/queue": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/QueueEntryDTO"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["QueueEntrys"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueueEntryDTO"
                                }
                            }
                        }
                    }
                },
                "tags": ["QueueEntrys"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueueEntryDTO"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueueEntryDTO"
                                }
                            }
                        }
                    }
                },
                "tags": ["QueueEntrys"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueueEntryDTO"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["QueueEntrys"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/offSchedule": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/EpisodeCacheDefinition"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeCacheDefinition"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeCacheDefinition"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeCacheDefinition"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeCacheDefinition"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/offSchedule/refresh": {
            "post": {
                "operationId": "Refresh",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeCacheDefinition"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/config": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Record_string.string_"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Settings"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record_string.string_"
                                }
                            }
                        }
                    }
                },
                "tags": ["Settings"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Record_string.string_"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record_string.string_"
                                }
                            }
                        }
                    }
                },
                "tags": ["Settings"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Record_string.string_"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["Settings"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/apps": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/App"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}